+++ server/src/SocketManager.cpp
@@ -10,7 +10,7 @@ void sock_info_t::init()
 	m_serv_addr.sin_port = htons(6666);
 	m_serv_addr.sin_addr.s_addr = htonl(INADDR_ANY);
 	m_listen_fd = socket(AF_INET, SOCK_STREAM, 0);
-	int opt = 1;
+	int16_t opt = 1;
 	setsockopt(m_listen_fd,SOL_SOCKET, SO_REUSEADDR, &opt, sizeof(opt));
 	if(bind(m_listen_fd, (struct sockaddr*)&m_serv_addr, sizeof(m_serv_addr)) == -1)
     {
